Allow to use Poké Balls out of battle
Updated ItemUseOutOfBattle_PokeBall I made it handle the access to the party screen. Now I can forget about using the Poke Balls' .type item data field for that purpose, which in turn means that the only change required for the Poké Balls' item data now is to assign the out of battle effect. By doing this, the branch can keep working on clean copies of Pokeemerald like normal, and it can also work in up-to-date copies of the Pokeemerald-expansion too.